Removed unneeded file, fixed react warnings
