+++ portal-ui/src/screens/Console/Users/DeleteUser.tsx
@@ -23,13 +23,17 @@ import ConfirmDialog from "../Common/ModalWrapper/ConfirmDialog";
 import { ErrorResponseHandler } from "../../../common/types";
 import { ConfirmDeleteIcon } from "../../../icons";
 import { encodeURLString } from "../../../common/utils";
+import WarningMessage from "../Common/WarningMessage/WarningMessage";
+import TableWrapper from "../Common/TableWrapper/TableWrapper";
 import api from "../../../common/api";
-
+import { IAM_PAGES } from "../../../common/SecureComponent/permissions";
+import Loader from "../Common/Loader/Loader";
 interface IDeleteUserProps {
   closeDeleteModalAndRefresh: (refresh: boolean) => void;
   deleteOpen: boolean;
   selectedUsers: string[] | null;
   setErrorSnackMessage: typeof setErrorSnackMessage;
+  history: any;
 }
 
 const DeleteUser = ({
@@ -37,6 +41,7 @@ const DeleteUser = ({
   deleteOpen,
   selectedUsers,
   setErrorSnackMessage,
+  history,
 }: IDeleteUserProps) => {
   const onDelSuccess = () => closeDeleteModalAndRefresh(true);
   const onDelError = (err: ErrorResponseHandler) => setErrorSnackMessage(err);
@@ -44,19 +49,29 @@ const DeleteUser = ({
 
   const [deleteLoading, invokeDeleteApi] = useApi(onDelSuccess, onDelError);
   const [loadingSA, setLoadingSA] = useState<boolean>(true);
+  const [hasSA, setHasSA] = useState<boolean>(false);
   const [userSAList, setUserSAList] = useState<userSACount[]>([]);
 
   const userLoggedIn = localStorage.getItem("userLoggedIn") || "";
+ 
+   useEffect(() => {  
 
-  useEffect(() => {
-    if(selectedUsers){
-    checkUserSA(selectedUsers)
-    };
-  }, []);
-
-  useEffect(() => {    
-    console.log("You set userSAList to:", userSAList)    
-  }, [userSAList]);
+     if(selectedUsers) {  
+      api
+      .invoke("POST", `/api/v1/users/service-accounts`, selectedUsers)
+      .then((res) => {  
+        setUserSAList(res.userServiceAccountList) ;      
+        if (res.hasSA) {
+          setHasSA(true)
+        }           
+        setLoadingSA(false);
+      })
+      .catch((err: ErrorResponseHandler) => {
+        setErrorSnackMessage(err);
+        setLoadingSA(false);
+      });   
+     } 
+   }, [selectedUsers, setErrorSnackMessage]);
 
   if (!selectedUsers) {
     return null;
@@ -66,6 +81,17 @@ const DeleteUser = ({
       <b>{user}</b>
     </div>
   ));  
+  const viewAction = (selectionElement: any): void => {
+    history.push(
+      `${IAM_PAGES.USERS}/${encodeURLString(selectionElement.userName)}`
+    );
+  };
+  const tableActions = [
+    {
+      type: "view",
+      onClick: viewAction,
+    },
+  ];
 
   const onConfirmDelete = () => {
     for (let user of selectedUsers) {
@@ -84,23 +110,15 @@ const DeleteUser = ({
   interface userSACount {
     userName: string;
     numSAs: number;
-  }
+  } 
 
-  const checkUserSA = (users : string[]) => {
-   api
-        .invoke("POST", `/api/v1/users/service-accounts`, users)
-        .then((res) => {  
-          console.log("In the getUserSACount api res.userServiceAccountList: ", res.userServiceAccountList)
-          setUserSAList(res.userServiceAccountList) ;                 
-          setLoadingSA(false);
-        })
-        .catch((err: ErrorResponseHandler) => {
-          setErrorSnackMessage(err);
-          setLoadingSA(false);
-        });
-  };
+  const noSAtext = "Are you sure you want to delete the following " + selectedUsers.length+" "+
+  "user"+ (selectedUsers.length > 1 ? "s?" : "?")
 
   return (
+    loadingSA ? 
+    <Loader />
+    :
     <ConfirmDialog
       title={`Delete User${selectedUsers.length > 1 ? "s" : ""}`}
       confirmText={"Delete"}
@@ -110,10 +128,32 @@ const DeleteUser = ({
       onConfirm={onConfirmDelete}
       onClose={onClose}
       confirmationContent={
-        <DialogContentText>
-          Are you sure you want to delete the following {selectedUsers.length}{" "}
-          user{selectedUsers.length > 1 ? "s?" : "?"}
-          <b>{renderUsers}</b>
+        <DialogContentText>  
+          
+        {hasSA ? 
+        <>
+          <WarningMessage 
+            label = "Click on a user to view the full listing of asociated Service Accounts. All Service Accounts associated with a user will be deleted along with the user. Are you sure you want to continue?"
+            title =  "Warning: One or more users selected has associated Service Accounts. "
+          />
+          <TableWrapper 
+            itemActions={tableActions}
+            columns={[
+              { label: "Username", elementKey: "userName" },
+              { label: "# Associated Service Accounts", elementKey: "numSAs" },
+            ]}
+            isLoading={loadingSA}
+            records={userSAList}
+            entityName="User Service Accounts"
+            idField="userName"
+            customPaperHeight="250"
+          />
+        </>
+        : <>
+            {noSAtext}
+            {renderUsers}
+          </>
+        }
         </DialogContentText>
       }
     />
@@ -127,7 +167,3 @@ const mapDispatchToProps = {
 const connector = connect(null, mapDispatchToProps);
 
 export default connector(DeleteUser);
-
-{/*
- 
-*/}
